MAJOR CHANGES
Move to using the same upstream StrongSwan as we use in dev/prod.
Also move the jitike containers to use Debian Stretch, also
similar to dev/prod.

With these changes, we can now test the full JITIKE in these
here containers in the same way as we do in dev/prod.

Signed-off-by: Kyle Mestery <mestery@mestery.com>